[ADAM-1580] Eliminate ADAMRDDFunctions classes.
Resolves #1580. Does not compile due to abstract classes not implemented.
