+++ adam-core/src/main/scala/org/bdgenomics/adam/rdd/GenomicRDD.scala
@@ -27,14 +27,21 @@ import htsjdk.variant.vcf.{
   VCFHeaderLineCount,
   VCFHeaderLineType
 }
+import org.apache.avro.Schema
+import org.apache.avro.file.DataFileWriter
 import org.apache.avro.generic.IndexedRecord
-import org.apache.hadoop.fs.{ FSDataOutputStream, FileSystem, Path }
+import org.apache.avro.specific.{ SpecificDatumWriter, SpecificRecordBase }
+import org.apache.hadoop.fs.{ FileSystem, Path }
+import org.apache.hadoop.mapreduce.{ OutputFormat => NewOutputFormat }
+import org.apache.parquet.avro.AvroParquetOutputFormat
+import org.apache.parquet.hadoop.ParquetOutputFormat
 import org.apache.parquet.hadoop.metadata.CompressionCodecName
-import org.apache.spark.SparkFiles
+import org.apache.parquet.hadoop.util.ContextUtil
+import org.apache.spark.{ SparkContext, SparkFiles }
 import org.apache.spark.api.java.JavaRDD
 import org.apache.spark.api.java.function.{ Function => JFunction, Function2 }
 import org.apache.spark.broadcast.Broadcast
-import org.apache.spark.rdd.RDD
+import org.apache.spark.rdd.{ InstrumentedOutputFormat, RDD }
 import org.apache.spark.sql.{ DataFrame, Dataset, SQLContext }
 import org.apache.spark.sql.functions._
 import org.apache.spark.storage.StorageLevel
@@ -55,7 +62,7 @@ import org.bdgenomics.formats.avro.{
 }
 import org.bdgenomics.utils.cli.SaveArgs
 import org.bdgenomics.utils.interval.array.IntervalArray
-import org.bdgenomics.utils.misc.Logging
+import org.bdgenomics.utils.misc.{ HadoopUtil, Logging }
 import org.json4s.JsonDSL._
 import org.json4s.jackson.JsonMethods._
 import scala.annotation.tailrec
@@ -126,6 +133,78 @@ trait GenomicRDD[T, U <: GenomicRDD[T, U]] extends Logging {
 
   override def toString = "%s with %d reference sequences"
     .format(getClass.getSimpleName, sequences.size)
+
+  /**
+   * Saves Avro data to a Hadoop file system.
+   *
+   * This method uses a SparkContext to identify our underlying file system,
+   * which we then save to.
+   *
+   * Frustratingly enough, although all records generated by the Avro IDL
+   * compiler have a static SCHEMA$ field, this field does not belong to
+   * the SpecificRecordBase abstract class, or the SpecificRecord interface.
+   * As such, we must force the user to pass in the schema.
+   *
+   * @tparam U The type of the specific record we are saving.
+   * @param pathName Path to save records to.
+   * @param sc SparkContext used for identifying underlying file system.
+   * @param schema Schema of records we are saving.
+   * @param avro Seq of records we are saving.
+   */
+  protected def saveAvro[U <: SpecificRecordBase](pathName: String,
+                                                  sc: SparkContext,
+                                                  schema: Schema,
+                                                  avro: Seq[U])(implicit tUag: ClassTag[U]) {
+
+    // get our current file system
+    val path = new Path(pathName)
+    val fs = path.getFileSystem(sc.hadoopConfiguration)
+
+    // get an output stream
+    val os = fs.create(path)
+
+    // set up avro for writing
+    val dw = new SpecificDatumWriter[U](schema)
+    val fw = new DataFileWriter[U](dw)
+    fw.create(schema, os)
+
+    // write all our records
+    avro.foreach(r => fw.append(r))
+
+    // close the file
+    fw.close()
+    os.close()
+  }
+
+  /**
+   * Saves an RDD to Parquet.
+   *
+   * @param args The output format configuration to use when saving the data.
+   */
+  def saveAsParquet(args: SaveArgs): Unit = {
+    saveAsParquet(args.outputPath,
+      args.blockSize,
+      args.pageSize,
+      args.compressionCodec,
+      args.disableDictionaryEncoding)
+  }
+
+  /**
+   * Saves an RDD of Avro data to Parquet.
+   *
+   * @param pathName The path to save the file to.
+   * @param blockSize The size in bytes of blocks to write.
+   * @param pageSize The size in bytes of pages to write.
+   * @param compressCodec The compression codec to apply to pages.
+   * @param disableDictionaryEncoding If false, dictionary encoding is used. If
+   *   true, delta encoding is used.
+   */
+  def saveAsParquet(
+    pathName: String,
+    blockSize: Int = 128 * 1024 * 1024,
+    pageSize: Int = 1 * 1024 * 1024,
+    compressCodec: CompressionCodecName = CompressionCodecName.GZIP,
+    disableDictionaryEncoding: Boolean = false): Unit
 
   /**
    * The RDD of genomic data that we are wrapping.
@@ -199,6 +278,21 @@ trait GenomicRDD[T, U <: GenomicRDD[T, U]] extends Logging {
    */
   lazy val jrdd: JavaRDD[T] = {
     rdd.toJavaRDD()
+  }
+
+  /**
+   * Save the sequence dictionary to disk.
+   *
+   * @param filePath The filepath where we will save the sequence dictionary.
+   */
+  protected def saveSequences(filePath: String): Unit = {
+    // convert sequence dictionary to avro form and save
+    val contigs = sequences.toAvro
+
+    saveAvro("%s/_seqdict.avro".format(filePath),
+      rdd.context,
+      Contig.SCHEMA$,
+      contigs)
   }
 
   /**
@@ -2136,6 +2230,19 @@ case class GenericGenomicRDD[T](
  */
 trait MultisampleGenomicRDD[T, U <: MultisampleGenomicRDD[T, U]] extends GenomicRDD[T, U] {
 
+  /**
+   * Save the samples to disk.
+   *
+   * @param filePath The filepath to the file where we will save the samples.
+   */
+  protected def saveSamples(filePath: String): Unit = {
+    // get file to write to
+    saveAvro("%s/_samples.avro".format(filePath),
+      rdd.context,
+      Sample.SCHEMA$,
+      samples)
+  }
+
   override def toString = "%s with %d reference sequences and %d samples"
     .format(getClass.getSimpleName, sequences.size, samples.size)
 
@@ -2745,19 +2852,6 @@ private[rdd] trait VCFSupportingGenomicRDD[T, U <: VCFSupportingGenomicRDD[T, U]
 abstract class MultisampleAvroGenomicRDD[T <% IndexedRecord: Manifest, U <: Product, V <: MultisampleAvroGenomicRDD[T, U, V]] extends AvroGenomicRDD[T, U, V]
     with MultisampleGenomicRDD[T, V] {
 
-  /**
-   * Save the samples to disk.
-   *
-   * @param filePath The filepath to the file where we will save the samples.
-   */
-  protected def saveSamples(filePath: String): Unit = {
-    // get file to write to
-    saveAvro("%s/_samples.avro".format(filePath),
-      rdd.context,
-      Sample.SCHEMA$,
-      samples)
-  }
-
   override protected def saveMetadata(filePath: String): Unit = {
     savePartitionMap(filePath)
     saveSequences(filePath)
@@ -2770,8 +2864,57 @@ abstract class MultisampleAvroGenomicRDD[T <% IndexedRecord: Manifest, U <: Prod
  * Avro IndexedRecords. This abstract class provides methods for saving to
  * Parquet, and provides hooks for writing the metadata.
  */
-abstract class AvroGenomicRDD[T <% IndexedRecord: Manifest, U <: Product, V <: AvroGenomicRDD[T, U, V]] extends ADAMRDDFunctions[T]
-    with GenomicDataset[T, U, V] {
+abstract class AvroGenomicRDD[T <% IndexedRecord: Manifest, U <: Product, V <: AvroGenomicRDD[T, U, V]] extends GenomicDataset[T, U, V] {
+
+  protected def saveRddAsParquet(args: SaveArgs): Unit = {
+    saveRddAsParquet(
+      args.outputPath,
+      args.blockSize,
+      args.pageSize,
+      args.compressionCodec,
+      args.disableDictionaryEncoding
+    )
+  }
+
+  /**
+   * Saves an RDD of Avro data to Parquet.
+   *
+   * @param pathName The path to save the file to.
+   * @param blockSize The size in bytes of blocks to write. Defaults to 128 * 1024 * 1024.
+   * @param pageSize The size in bytes of pages to write. Defaults to 1 * 1024 * 1024.
+   * @param compressCodec The compression codec to apply to pages. Defaults to CompressionCodecName.GZIP.
+   * @param disableDictionaryEncoding If false, dictionary encoding is used. If
+   *   true, delta encoding is used. Defaults to false.
+   * @param optSchema The optional schema to set. Defaults to None.
+   */
+  protected def saveRddAsParquet(
+    pathName: String,
+    blockSize: Int = 128 * 1024 * 1024,
+    pageSize: Int = 1 * 1024 * 1024,
+    compressCodec: CompressionCodecName = CompressionCodecName.GZIP,
+    disableDictionaryEncoding: Boolean = false,
+    optSchema: Option[Schema] = None): Unit = SaveAsADAM.time {
+    log.info("Saving data in ADAM format")
+
+    val job = HadoopUtil.newJob(rdd.context)
+    ParquetOutputFormat.setCompression(job, compressCodec)
+    ParquetOutputFormat.setEnableDictionary(job, !disableDictionaryEncoding)
+    ParquetOutputFormat.setBlockSize(job, blockSize)
+    ParquetOutputFormat.setPageSize(job, pageSize)
+    AvroParquetOutputFormat.setSchema(
+      job,
+      optSchema.getOrElse(manifest[T].runtimeClass.asInstanceOf[Class[T]].newInstance().getSchema)
+    )
+
+    // Add the Void Key
+    val recordToSave = rdd.map(p => (null, p))
+    // Save the values to the ADAM/Parquet file
+    recordToSave.saveAsNewAPIHadoopFile(
+      pathName,
+      classOf[java.lang.Void], manifest[T].runtimeClass.asInstanceOf[Class[T]], classOf[InstrumentedADAMAvroParquetOutputFormat],
+      ContextUtil.getConfiguration(job)
+    )
+  }
 
   /**
    * Save the partition map to disk. This is done by adding the partition
@@ -2804,21 +2947,6 @@ abstract class AvroGenomicRDD[T <% IndexedRecord: Manifest, U <: Product, V <: A
   }
 
   /**
-   * Save the sequence dictionary to disk.
-   *
-   * @param filePath The filepath where we will save the sequence dictionary.
-   */
-  protected def saveSequences(filePath: String): Unit = {
-    // convert sequence dictionary to avro form and save
-    val contigs = sequences.toAvro
-
-    saveAvro("%s/_seqdict.avro".format(filePath),
-      rdd.context,
-      Contig.SCHEMA$,
-      contigs)
-  }
-
-  /**
    * Called in saveAsParquet after saving RDD to Parquet to save metadata.
    *
    * Writes any necessary metadata to disk. If not overridden, writes the
@@ -2829,26 +2957,6 @@ abstract class AvroGenomicRDD[T <% IndexedRecord: Manifest, U <: Product, V <: A
   protected def saveMetadata(filePath: String): Unit = {
     savePartitionMap(filePath)
     saveSequences(filePath)
-  }
-
-  /**
-   * Saves RDD as a directory of Parquet files.
-   *
-   * The RDD is written as a directory of Parquet files, with
-   * Parquet configuration described by the input param args.
-   * The provided sequence dictionary is written at args.outputPath/_seqdict.avro
-   * as Avro binary.
-   *
-   * @param args Save configuration arguments.
-   */
-  def saveAsParquet(args: SaveArgs) {
-    saveAsParquet(
-      args.outputPath,
-      args.blockSize,
-      args.pageSize,
-      args.compressionCodec,
-      args.disableDictionaryEncoding
-    )
   }
 
   /**
@@ -2945,4 +3053,9 @@ abstract class AvroGenomicRDD[T <% IndexedRecord: Manifest, U <: Product, V <: A
     saveMetadata(filePath)
   }
 
+}
+
+private[rdd] class InstrumentedADAMAvroParquetOutputFormat extends InstrumentedOutputFormat[Void, IndexedRecord] {
+  override def outputFormatClass(): Class[_ <: NewOutputFormat[Void, IndexedRecord]] = classOf[AvroParquetOutputFormat[IndexedRecord]]
+  override def timerName(): String = WriteADAMRecord.timerName
 }
